# clients/go/api/openapi.yaml
openapi: 3.0.0
info:
  description: "API REST gamificada para gerenciamento de tarefas com IA, XP, levels\
    \ e sistema de recompensas"
  title: Gamified Task Management API
  version: 1.0.0
servers:
- description: Servidor de Desenvolvimento
  url: http://localhost:3000
security:
- bearerAuth: []
...
paths:
...
  /api/users/stats:
    get:
      operationId: getUserStats
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getUserStats_200_response'
          description: Estatísticas do sistema
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado - privilégios de admin necessários
      summary: Obter estatísticas gerais do sistema (Admin only)
      tags:
      - Users
...
  /api/users/{id}:
    delete:
      operationId: deleteUser
      parameters:
      - description: ID do usuário
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/deleteUser_200_response'
          description: Usuário deletado com sucesso
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Não é possível deletar própria conta
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado - privilégios de admin necessários
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Usuário não encontrado
      summary: Deletar usuário (Admin only)
      tags:
      - Users
    get:
      operationId: getUserById
      parameters:
      - description: ID do usuário
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getUserById_200_response'
          description: Detalhes do usuário
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado - privilégios de admin necessários
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Usuário não encontrado
      summary: Obter usuário específico (Admin only)
      tags:
      - Users
    put:
      operationId: updateUser
      parameters:
      - description: ID do usuário
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateUser_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/updateUser_200_response'
          description: Usuário atualizado com sucesso
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Email já está em uso
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado - privilégios de admin necessários
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Usuário não encontrado
      summary: Atualizar usuário (Admin only)
      tags:
      - Users
  /api/users/{id}/toggle-status:
    patch:
      operationId: toggleUserStatus
      parameters:
      - description: ID do usuário
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/toggleUserStatus_200_response'
          description: Status do usuário alterado com sucesso
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Não é possível desativar própria conta
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado - privilégios de admin necessários
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Usuário não encontrado
      summary: Ativar/Desativar usuário (Admin only)
      tags:
      - Users
components:
  schemas:
    User:
      example:
        gamification:
          totalXP: 450
          level: 5
        lastLogin: 2000-01-23T04:56:07.000+00:00
        role: user
        name: Yuri Dev
        id: 64a1b2c3d4e5f6789abcdef0
        isActive: true
        email: yuri.dev@example.com
      properties:
        id:
          example: 64a1b2c3d4e5f6789abcdef0
          type: string
        name:
          example: Yuri Dev
          type: string
        email:
          example: yuri.dev@example.com
          type: string
        role:
          enum:
          - user
          - admin
          example: user
          type: string
        isActive:
          example: true
          type: boolean
        lastLogin:
          format: date-time
          type: string
        gamification:
          $ref: '#/components/schemas/User_gamification'
      type: object
...
    Error:
      example:
        success: false
        message: Erro de validação
      properties:
        success:
          example: false
          type: boolean
        message:
          example: Erro de validação
          type: string
      type: object
...
    getUserProfile_200_response_data:
      example:
        user:
          gamification:
            totalXP: 450
            level: 5
          lastLogin: 2000-01-23T04:56:07.000+00:00
          role: user
          name: Yuri Dev
          id: 64a1b2c3d4e5f6789abcdef0
          isActive: true
          email: yuri.dev@example.com
      properties:
        user:
          $ref: '#/components/schemas/User'
      type: object
...
    getUserStats_200_response_data_users:
      example:
        total: 150
        inactive: 10
        active: 140
        admins: 5
      properties:
        total:
          example: 150
          type: integer
        active:
          example: 140
          type: integer
        inactive:
          example: 10
          type: integer
        admins:
          example: 5
          type: integer
      type: object
    getUserStats_200_response_data_tasks:
      example:
        total: 1250
        inProgress: 150
        overdue: 45
        pending: 300
        completed: 800
      properties:
        total:
          example: 1250
          type: integer
        completed:
          example: 800
          type: integer
        pending:
          example: 300
          type: integer
        inProgress:
          example: 150
          type: integer
        overdue:
          example: 45
          type: integer
      type: object
    getUserStats_200_response_data_mostActiveUsers_inner:
      example:
        taskCount: 0
        name: name
        userId: userId
        email: email
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
        taskCount:
          type: integer
      type: object
    getUserStats_200_response_data:
      example:
        mostActiveUsers:
        - taskCount: 0
          name: name
          userId: userId
          email: email
        - taskCount: 0
          name: name
          userId: userId
          email: email
        users:
          total: 150
          inactive: 10
          active: 140
          admins: 5
        tasks:
          total: 1250
          inProgress: 150
          overdue: 45
          pending: 300
          completed: 800
      properties:
        users:
          $ref: '#/components/schemas/getUserStats_200_response_data_users'
        tasks:
          $ref: '#/components/schemas/getUserStats_200_response_data_tasks'
        mostActiveUsers:
          items:
            $ref: '#/components/schemas/getUserStats_200_response_data_mostActiveUsers_inner'
          type: array
      type: object
    getUserStats_200_response:
      example:
        data:
          mostActiveUsers:
          - taskCount: 0
            name: name
            userId: userId
            email: email
          - taskCount: 0
            name: name
            userId: userId
            email: email
          users:
            total: 150
            inactive: 10
            active: 140
            admins: 5
          tasks:
            total: 1250
            inProgress: 150
            overdue: 45
            pending: 300
            completed: 800
        success: true
      properties:
        success:
          example: true
          type: boolean
        data:
          $ref: '#/components/schemas/getUserStats_200_response_data'
      type: object
...
    getUserById_200_response_data_taskStats:
      example:
        total: 0
        in_progress: 1
        pending: 6
        completed: 5
      properties:
        total:
          type: integer
        pending:
          type: integer
        in_progress:
          type: integer
        completed:
          type: integer
      type: object
    getUserById_200_response_data:
      example:
        taskStats:
          total: 0
          in_progress: 1
          pending: 6
          completed: 5
        user:
          gamification:
            totalXP: 450
            level: 5
          lastLogin: 2000-01-23T04:56:07.000+00:00
          role: user
          name: Yuri Dev
          id: 64a1b2c3d4e5f6789abcdef0
          isActive: true
          email: yuri.dev@example.com
      properties:
        user:
          $ref: '#/components/schemas/User'
        taskStats:
          $ref: '#/components/schemas/getUserById_200_response_data_taskStats'
      type: object
    getUserById_200_response:
      example:
        data:
          taskStats:
            total: 0
            in_progress: 1
            pending: 6
            completed: 5
          user:
            gamification:
              totalXP: 450
              level: 5
            lastLogin: 2000-01-23T04:56:07.000+00:00
            role: user
            name: Yuri Dev
            id: 64a1b2c3d4e5f6789abcdef0
            isActive: true
            email: yuri.dev@example.com
        success: true
      properties:
        success:
          example: true
          type: boolean
        data:
          $ref: '#/components/schemas/getUserById_200_response_data'
      type: object
    updateUser_request:
      properties:
        name:
          example: João Silva Santos
          maxLength: 50
          minLength: 2
          type: string
        email:
          example: joao.santos@email.com
          format: email
          type: string
        role:
          enum:
          - user
          - admin
          example: user
          type: string
        isActive:
          example: true
          type: boolean
      type: object
    updateUser_200_response:
      example:
        data:
          user:
            gamification:
              totalXP: 450
              level: 5
            lastLogin: 2000-01-23T04:56:07.000+00:00
            role: user
            name: Yuri Dev
            id: 64a1b2c3d4e5f6789abcdef0
            isActive: true
            email: yuri.dev@example.com
        success: true
        message: Usuário atualizado com sucesso
      properties:
        success:
          example: true
          type: boolean
        message:
          example: Usuário atualizado com sucesso
          type: string
        data:
          $ref: '#/components/schemas/getUserProfile_200_response_data'
      type: object
    deleteUser_200_response:
      example:
        success: true
        message: Usuário e suas tarefas foram deletados com sucesso
      properties:
        success:
          example: true
          type: boolean
        message:
          example: Usuário e suas tarefas foram deletados com sucesso
          type: string
      type: object
    toggleUserStatus_200_response_data_user:
      example:
        name: name
        id: id
        isActive: true
        email: email
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        isActive:
          type: boolean
      type: object
    toggleUserStatus_200_response_data:
      example:
        user:
          name: name
          id: id
          isActive: true
          email: email
      properties:
        user:
          $ref: '#/components/schemas/toggleUserStatus_200_response_data_user'
      type: object
    toggleUserStatus_200_response:
      example:
        data:
          user:
            name: name
            id: id
            isActive: true
            email: email
        success: true
        message: Usuário ativado com sucesso
      properties:
        success:
          example: true
          type: boolean
        message:
          example: Usuário ativado com sucesso
          type: string
        data:
          $ref: '#/components/schemas/toggleUserStatus_200_response_data'
      type: object
    User_gamification:
      example:
        totalXP: 450
        level: 5
      properties:
        level:
          example: 5
          type: number
        totalXP:
          example: 450
          type: number
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http